attempt to simplify left/width calendar logic
