+++ src/components/content/base.ts
@@ -1,20 +1,18 @@
 import type { Dayjs } from "dayjs";
 
 export interface Event {
-	startRelative?: string;
-	start: string;
-	end: string;
-	overlappingDays: Event[];
+	date: Dayjs;
+	start: string | Dayjs;
+	end: string | Dayjs;
+
 	name: string;
-	timed: boolean;
 	backgroundImage?: string;
 	description: string;
-	height: number;
-	top: number;
-	leftPercentage: number;
-	widthPercentage: number;
-	zIndex?: 'auto' | number;
-	relative: Dayjs;
+
+	top?: number;
+	height?: number;
+	leftPercentage?: number;
+	widthPercentage?: number;
 };
 
 export function slugify(s: string) {
